feat: implement file and redis caching mechanisms
